refactor : repackage product
상품 관련 패키지 구조 변경
